refactor: add content seed data
Add knex seed data file which contains a partial dump of the following tables:
* posts
* publications
* tags
* tags_count

In addition, it generates random time, up to one month before the time of running the see, for each post.
This makes sure that most posts are "new" and will be part of the latest feed.

Closes #11