feat: bookmark list
Bookmarks now can be organized in lists. Each bookmark can belong to one list only.
Each list has a name and the user can create, remove and rename a bookmark list.
When a list is deleted the bookmarks are ejected and go back to the default list.
Bookmark list is a premium feature and not available to everyone