refactor: denormalize tags to post table
Fetching the tags for each post in a feed may affect on the feed performance.
To reduce this overhead at insert time the top 5 tags will be inserted to the post object itself.
When fetching the feed there is no need to join tables and order tags by popularity any more.