refactor: pagination optimization
To boost performance, I reimplemented the pagination mechanism to use `where` instead of `offset`.
`where` clause can utilize indexes and increase drastically the performance of the query compared to the traditional `offset` implementation.
This implementation is backward compatible thanks to the cursor pagination that is already implemented in the client