feat: public bookmarks with rss support
This PR adds support in sharing bookmarks using RSS.
It adds a new GraphQL query for receiving the sharing settings, `bookmarksSharing`. A mutation fo manipulating the settings `setBookmarksSharing`. And revives the old RSS functionality that we had for the premium users with a slight change.

DD-515 #done
DD-516 #done
DD-518 #done