refactor: user reading rank with tags (#872)
DD-503 #done

As part of DailyFest working on the Weekly Goals 2.0 - there are certain APIs that need to be updated based on the new requirements. With that, the APIs are backward compatible whenever it misses the necessary parameters such as version or date range params.

Date range parameters are now acceptable on most APIs to support flexible time frames.