fix: post import seeds for development (#926)
DD-594 #done

Updating the canonical URL's index to be unique made the existing seeds to be invalid since the updating of existing data as part of the actual migration. We now updated the seeds by going back to the commit before and exporting the result after running the migration for applying the said constraint.