fix: user mention beside a special character (#969)
DD-582 #done

Fixes an issue where the user has commented beside a special character, either a comma, period, or whatnot. This PR allows the support like how it is on Twitter when a mention is beside a special character. We cut all the words by space then once we filter each word for a mention, we replace all special characters with a space and map the replaced value in an array. We will then split the word by their replaced space and check if any mention was made.