refactor: author emails to send on scout (#993)
WT-247 #done

There are multiple emails we send to authors, two of which are when a user has commented on a post or someone made a comment on a comment. We should be sending those notifications on scout as well. This includes the analytics we have planned to be sent to authors, and are now to be sent to scouts.