refactor: removal of importing keyword and only as part of migration (#1057)
There was a recent introduction in one of the migration files where it would also automatically populate the `Keywords` table. With that, running import command results in an error and requires manual intervention. To make it easier for starting a clean slate, we decided to remove importing the `Keywords` from the command itself and let the migration do the work.