feat: add authentication support for graphql subscriptions
Cookies are not provided when using WebSocket in the extension. Hence, our authentication doesn't work in GraphQL Subscriptions from the extension.
This PR adds another step in the connection to explicitly check the token provided by the client. (A PR for the apps will be created shortly)
This is required for the notifications