feat: generate notifications (#1087)
Generate all supported notifications using background workers.
This PR introduces new concepts:
* `NotificationWorker` - Lightweight variation of our good old `Worker`
that its sole purpose is to generate notifications. It returns an array
of notification types to generate along with their relevant context
(parameters).
* `NotificationBuilder` - To make it easy to build new notification
types, the builder has pre-built functions for common use cases, and
everything is chainable to make it more readable.
* `NotificationBaseContext` - Used to store all the parameters a single
notification needs in order to be generated completely. This type has
inherited types that makes it flexible to provide only the needed
parameters per type.

The `generateNotificationMap` object holds a map that using the
`NotificationBuilder` it constructs the notification object that later
on can be stored.

The storage mechanism is exactly-once, meaning that if for some reason
pub/sub decided to replay the message (which can happen), we will still
save it only once and there will be no notification spam.

It's a huge PR, I suggest starting with the notifications folder as it
contains all the fundamental stuff and then you can review the new
workers.

WT-786 #done