refactor: send emails based on notification events (#1090)
Introducing a change in the way we send notification emails! 🤩 
Previously, we subscribed to many events that trigger an email. Now we
have one subscriber that listens to new notifications and, based on that
decides what email to send and to whom. A consolidation of all the
emails sent to a single file.

This change also uses a new version of emails some of them are backward
compatible but some are not. The following emails will have to be
activated only once we merge this PR:
* Report post - Approved
* Source Request - Approved
* Notification - Reply
* Notification - Upvotes

WT-781 #done