fix: add contacts to sendgrid
For the past few months there was an issue with adding registered users to sendgrid contact lists.
I added a new worker that subscribes to new users event and if the info is confirmed, it will add it to sendgrid. Before that we listened only for updates, which caused many users not to be added.
Also when we checked for updates the condition we had there didn't support the case where the only property changed is the info confirmed.