feat: action table (#1251)
Introduction of `action` table. The workers and the retro checks will be
worked in a different PR so we can move forward with the tasks that
require this information.

Thoughts: the whole schema for action is pretty small, would you rather
have it in the `users` schema or continue with having its own?