feat: subscribe to notification type of notification (#1255)
Introduced a new type of notification `squad_subscribe_to_notification`.
Have added the prefix `squad` to the type since the content has a
context of the squad and most likely we will have another
`subscribe_to_notification` type.

I was planning to use the OneSignal API/SDK to check whether the user
has any subscribed devices, but after checking, the said feature is not
yet available. So now, it would depend on the FE whether we will render
this notification item based on their subscription.