feat: we should auto populate the usernames for auth onboarding (#1256)
Created a new query `generateUniqueUsername `, that takes the users name
on signup and attempts to generate a username for them. It has 3
attempts to do so or it will exit without a suggestion. If the name of
the user is already a username in the system, it will attempted 3 times
to add a random string of numbers at the end to generate a unique
username

WT-804 #done

---------

Co-authored-by: Chris Bongers <chrisbongers@gmail.com>