fix: flaky test due to order (#1293)
The score value of the order by the score is the same, the query
sometimes returns `p1` first over the other and vice-versa. Since they
have the same score in terms of the actual issue, there should be no
concern with regard to which one should come first or second.