feat: use growthbook as an experimentation platform (#1356)
Add support for growthbook.io as experimentation and feature flags platform. 
This will replace flagsmith and will allow us much more flexible experiments.

Boot now includes a new object `exp` that contains the recent active experiments of the user (as `e`) and encrypted features settings (as `f`).

The frontend uses it to decide what values to value to assign the feature.