feat: muting notification (#1353)
We created a map that defines which type of `NotificationPreference` a
`NotificationType` belongs to. This is to understand whether we should
put the `uniqueKey` on the designated properties unique for each entity.

It was mentioned to include a mutation to `unmute` a preference. I
thought of making it a more general kind of unmuting since its sole
purpose would be to remove the actual row, I have instead renamed it a
bit to be applicable to removing `subscribed` in the future.