fix: muting squad new comment (#1367)
Just like the `CommentReply` having its counterpart as `SquadReply`, it
goes the same with `ArticleNewComment` having `SquadNewComment`.

By including the `SquadNewComment` in the list, we are now allowing it
to be muted based on the checks made by its mutation. Have also added a
test to ensure the worker recognizes the muting capability.