fix: correct notification preference type (#1370)
after further development, I noticed that even though we strictly set
the proper `type` of the `NotificationPreference`, it is getting
overridden by what was defined in the class. Using the base class
results in always setting it to `NotificationPreference` string as shown
below. I have pushed a commit to fix this and ended up with a switch
case, I tried doing a map of the class but wasn't working.