infra: ignore changes to views and discussion score
We have cron jobs that update frequently the views and discussion score of posts. These changes generate quite a lot of cdc messages but we ignore them on our cdc worker. This change ignores them on debezium level to reduce the overall load on the system.

Please extra check my conditions!