feat: experimentation support in digest
Add a variation column to the digest table.
Using the variation we will be able to experiment with different email settings and parameters.

It's still WIP as product haven't concluded the exact experiment, but they are certain they want to set the title to the post title and change the sender.