fix: id conflicts when signing up
Due to legacy reasons we generate user id upon first visit and this can cause quite a lot of issues.
One of them is id clashes on sign up. This PR handles the id clashes and retries several times before giving up.
I also refactored a bit to make it look better 😬

It's not ready to be merged yet as I wait for an heimdall change that will be deployed tomorrow hopefully.