chore: enable cron for referral reminder (#1556)
The count of users with `false` for `showGenericReferral` is 77k. If it
goes down, it means the cron has updated users mistakenly since everyone
has either the `true` value of the said property or has seen it.

Again, thank you to @rebelchris for running the scripts on a day-to-day
basis 🥳

Note: kindly ignore the changes on other lines, made by prettier
automatically.