feat: get comment by id query (#1553)
In order to lessen the load we request from the server, we will start to
not fetch the markdown content on post comments as those values are
unused. If need be, the app will send a query to fetch the comments
details which include what was missed in the post comments, which is
the markdown content.