refactor: type safety pubsub implementation
**This PR is a proposal nothing that should go to production at the moment.**

The main goal is to ensure type safety when using pubsub both on the publisher and consumer side and keep them consistent.

My approach uses a centralized schema type that is enforced by a single function for sending events (`triggerTypedEvent`) and on the worker side by a new `TypedWorker` interface.
The DX is significantly improved as you get auto complete and validation for topic names and payload as long as we have the schema up to date.

In this PR, I migrated some topics and a single worker to show how we can make it work if we agree on this approach.
I wanted to demo how it's gonna look before suggesting an action plan so if you don't like it at all, say it explicitly.