feat: feed side caching support (#1579)
We gonna run an internal test with v20 to move the responsibility of the
caching to the feed side.

This means we have to respect the cursor returned by the feed and also
provide it on every call.

I hacked my way around for the testing and once we confirm and I can
remove all the legacy stuff, I'll make it prettier. I promise.