feat: notification v2 email worker
Same drill, this worker is not connected to production. The next PR will cover that.

As mentioned in the DR, I utilized the `personalizations` that allows sending multiple emails in one sendgrid call. Pretty rad!

There was lots of good old copy-paste because v1 worker is going to be deleted soon.

I also copied the tests and adjusted them accordingly.

WT-2045