refactor: notification attachment to support video (#1605)
The nature of our notification attachment is, after generating it, the
entity won’t be attached to the relevant entity anymore. Meaning, when
rendering it, we won’t know whether the relevant entity came from a
video in order for us to add a play button overlay. We need to have a
column to tell us that the attachment was from a video-related entity.
---------

Co-authored-by: capJavert <ante@kickass.website>