feat: automation infra (#1666)
New generic endpoint for working with automation services (e.g Retool).

The idea is to provide as much flexibility as possible so it just passes
through any parameters from the client. It adds the `userId` to any call
and also checks for authentication.

This way we just need to add new automation to the object and viola!

I also have some ideas on how to store the automations in the DB, but
that's too early for that.