feat: reading reminder support for digest system
As we want to add a daily reminder for users to read a post, I decided to use the existing daily digest system. 💪

This PR extends the digest system and allows us to leverage it for different use cases.

This is a proposal as the RFC wasn't approved yet. So feel free to reject this idea.