feat: set watercooler to be opt-in (#1865)
Make watercooler as opt-in even though it's a public feed. Watercooler
mostly has memes which not all the users want, surfacing it as default
to all users seems to be controversial.

I also fixed a bug that even though "show on my feed" was set to false
there could be an edge case where we show the posts of the squad.
Unrelated to the original task but I noticed it as I go.