chore: reduce redis instance size (#1862)
Last week I deleted ~4 GiB worth of old and untouched Redis keys. At the
time we were using 5 out of 10 GiB of memory, which had been stable for
months. Now we use 1.3 out of 10 GiB, so we can reduce it to 3 GiB and
disable read-replica as we're not using it.


![image](https://github.com/dailydotdev/daily-api/assets/1681525/3cd9955e-f4ad-4f6d-986a-a3f1c5f73cb8)