fix: streak reset over the weekend (#2214)
I've added three more tests that involve the weekend to solidify the
integrity of the process. Found another set of edge cases along the way,
but should have been fixed by this PR now.

The edge cases were about, when it is the second day of the week, and
the user lost their streak, their last read could've been Friday,
Saturday, or Sunday. Have added test cases to each.

Also, to make it easier to maintain, I moved the functions to a separate
file.