fix: setting of cache when timezone is involved (#2231)
Most of our streaks data are queried and the correct timezones are
fetched. However, in CDC, the data being passed is in UTC, hence, in JS
we need to convert it to the right timezone for a proper calculation.

Added to tests:
- Timezone > UTC
- Timezone < UTC

Validated the entire CDC process relating to the setting of cache, every
end considers timezone now.