feat: order by priority column (#2260)
- Introduce a new column to handle ordering based on our demand.
- Apparently, there are some changes with the categories. Rather than
reusing an existing entity and just updating the title, it might become
inaccurate. Hence, deleting the rows instead. Good thing that we also
have no Squads with any category yet. Although they should be safe as
cascading would take effect and null would be set in any case.
  - Another option is to turn the `enabled` to `false` values.
  - ~~Removed `['General', 'DevOps', 'Cloud', 'Data']`.~~
- ~~Added `['Basics', 'DevOps & Cloud', 'AI', 'Open Source',
'DevRel']`~~
- Removed every category (good thing no one is using any of the items
yet). This is to make it easier to manage in the future and better
devex. We will run a once off script to insert the data along with the
Squad updates. This would make seeding our local env consistent.
- Created the migration for the column and did the necessary
insert/delete/update commands at the same time.