refactor: cache expiry to be next weekday (#2317)
We set the expiry of the cache until the next day, regardless if the
next day is a weekend.

There is one issue though, we have a cron job that resets the cache in
an interval, and when it runs on the weekend, let's say Saturday, the
cache expires on Sunday which the user might have not logged in and
defeats the purpose of the restore on weekends.

We are now setting the cache until the next valid day, so the previous
case still applies on weekdays, but now, on weekends, we set the expiry
until EOD of the next weekday. In the case above, EOD Monday.