refactor: use a single query for feed filters
In an effort to reduce the load on the database, I merged together the different queries used to generate the feed filters object.

It also includes some minor optimizations.

If this will not reduce the load we can consider redis. Now with a single query it can be easier to cache it. I may also consider moving the javascript logic to the query itself but let's see.