fix: external links submission (#2428)
There are two things fixed here, separated by each commit:
- In our CDC pipeline, there can be a case where two of the same links
were submitted. Is it not possible to process them at the same time when
approved, in this case, the external link would have been created on the
first instance, and on the second instance, it would produce an error
due to having the same link.
- Second fix was on the submission of post for moderation. The values
passed through the mutation would be present depending on the case. We
have title, content, and commentary but the post entity only has a title
and content.
- Freeform: must have a title and content. Empty `titleHtml` but
`contentHtml` should be defined.
- Share (from internal id): must have `title` and `titleHtml`. Empty
`content` and `contentHtml`.
- Share (from external link): must have `title` that is going to be used
as the title of the then-created Article post, which won't have
`titleHtml`. The `content` and `contentHtml` should be present and be
used as the `title` of the `Share` post created.
- Also made the tests more explicit.