fix: streak issue - fetching last recover date with timezone (#2452)
This is a 2-part PR series about the issues surrounding the streak
reset.

In this PR, this fixes an issue where the user has just restored their
streak, but suddenly it was broken again. Referring to the issue
mentioned here:
https://github.com/dailydotdev/daily/issues/1538#issuecomment-2462438066

The screenshot below tells the user did not read for two days straight,
however, on the next screenshot, it can be seen that the user restored
the streak that should allow him to continue the streak even when 2 days
had passed.
![Screenshot 2024-11-16 at 11 06
19 AM](https://github.com/user-attachments/assets/d70067d1-1409-43b5-ad2a-e163f6face18)

For context, the user lost his streak and restored it as seen below. At
12:03 the email notification about the streak being reset was sent, and
immediately after, the restoration happened.
![Screenshot 2024-11-16 at 10 54
24 AM](https://github.com/user-attachments/assets/49dd4225-cb8a-447c-9a4d-53d50f973b2c)

We have a cron job that runs on an internal, and it takes consideration
of the current streak and restores performed by the user. However, there
was an issue with how we pulled the data about the date of recovering
the streak.

In postgresql, setting a timezone on a `timestamp` column (without
timezone information column), would either push or pull the time
depending on the timezone value. In order to set correctly the timezone,
we must convert the value itself to a `timestamptz`, and then set the
correct timezone preferred by the user.

This explains the reset that happened, the query to pull all users that
should get reset was inaccurate by converting inaccurate value from the
recovery date.

Note: to show that it actually fixes the issue, I've introduced a test
with the user's configuration. I will [revert the fix
temporarily](https://github.com/dailydotdev/daily-api/pull/2452/commits/29442290ead71a2b65094f77cb0dc0a2c3fce88e)
to see if the test would fail.