refactor: create helpers to make funnel boot per use case
As we are about to swap onboarding to use funnels, I thought it can be a good idea to put the building blocks to help with that. To make sure we follow all the best practices.

I also fixed an issue that was recovering a session even though it shouldn't