feat: introduce pre-insert and pre-update hooks to posts
Initially I started introducing a block by words in title feature, and it turned out to be a mess. So I decided to introduce hooks that will allow us to easily do that and also expand in the future. As part of the initial setup, I also do the vordr check there as well as added title check