feat: dedup key for posts
The new deduplication key helps the feed to well, dedup posts.
We already have a solution place for shared post on feed side, but this is a more generic solution that supports also freeform.

For freeform we generate a hash so we can get rid of all spams

BONUS: CLAUDE.md file 😬